Reading and parsing of input file
The input file is csv formatted with a single header line describing the
ingredients, the other lines being rows of ingredient measures. The header may
look like this:

Flour, Eggs, Sugar

And the measurement rows like this:

1 cup, 150 grams, 0
1.5 cups, .1kg,2 tbsp

As you can see, weight and volume units may be mixed within the same row. If
the measure is zero, no unit is needed otherwise it is mandatory. Unit synonyms
may be used (so 'gram', 'grams' or 'g' are equally valid). There need not be a
space between the measurement value and unit. Uppercase or lowercase is fine.
The leading zero in a floating point value may be omitted, so 0.1g is the same
as .1g

A great deal of freedom has been given in the formatting since the information
is currently entered by hand. You could say that it has been optimized for
fewer key strokes and an amount of sloppiness.

If more efficient methods of data entry are added, this format could be
tightened up.