Merge ingredient columns
Merge and combine whole or partial columns. To allow for the fact that
ingredients are sometimes substituted, but the recipes to be generated from
the data will be more easily understood and compared when the ingredients are,
somehow, recombined.

A difficulty arises when two ingredients have, for example, different fat
content. Imagine we have three columns:

Butter, Milk, Cream

If we want to combine the milk and cream columns to make a single 'liquid'
column, we can transfer the fat content (20-40%) of the cream column to the
butter column and the rest to the milk column.

Note: that there is currently no support for column relabeling so the
merged column will be called 'milk' or 'cream', but not 'liquid' in the example
above.